fix(parity): byte->int8 sign-extension class (audit Group 4, Theme B)
Policy decision (confirmed with maintainer): convert Java `byte` fields to Go
int8 so int(field) sign-extends like Java's byte->int promotion, rather than
zero-extending via []byte. Perf-neutral (both 1 byte; sign vs zero extend is the
same single instruction), removes the latent trap, matches project convention.

- PixFont.CharMask [][]byte -> [][]int8 (MEDIUM, gfx-pix32-pixfont #10): the
  advance-trim sums `var12 += int(CharMask[i][...])` (DrawChar setup) now
  sign-extend like Java (PixFont.java:78,87). Mask pixel reads only test `== 0`,
  so signedness is irrelevant there; widened DrawChar/DrawMask/DrawCharAlpha/
  DrawMaskAlpha mask params to []int8. G1B() returns int8, so the byte() casts
  at unpack drop out.

- WordFilter.BadCombinations [][][]byte -> [][][]int8 (LOW, wordenc): ComboMatches
  and GetIndex now operate on signed int8, matching Java comboMatches(byte,...)'s
  signed binary-search ordering (WordFilter.java:679). Inert for real data
  (combos are getIndex codes 1..38) but bug-for-bug exact for bytes >127.

- Wall/Decor/GroundDecor/Location.Info byte -> int8 (INFO->convention): these are
  always read as int(Info)&0xFF, so behavior is unchanged; converted for
  convention consistency. Contained: setters keep byte params and cast int8(...)
  at assignment (Java's (byte) reinterpret); masked reads rewritten to
  int(Info)&0xFF (int8&0xFF won't compile).

+++ pkg/jagex2/dash3d/world3d/world3d.go
@@ -246,7 +246,7 @@ func (w *World3D) AddGroundDecoration(arg0 *model.Model, arg2, arg3, arg4, arg5 
 	var9.Z = arg4*128 + 64
 	var9.Y = arg7
 	var9.BitSet = arg3
-	var9.Info = arg6
+	var9.Info = int8(arg6) // Java: byte field; (byte) reinterpret of the typecode
 	if w.LevelTiles[arg5][arg2][arg4] == nil {
 		w.LevelTiles[arg5][arg2][arg4] = typ.NewGround(arg5, arg2, arg4)
 	}
@@ -285,7 +285,7 @@ func (w *World3D) AddWall(angle2, y, level, angle1 int, model1 *model.Model, mod
 
 	wall := typ.NewWall()
 	wall.BitSet = typecode1
-	wall.Info = typecode2
+	wall.Info = int8(typecode2)
 	wall.X = tileX*128 + 64
 	wall.Z = tileZ*128 + 64
 	wall.Y = y
@@ -311,7 +311,7 @@ func (w *World3D) SetWallDecoration(y, z, zOffset, typecode, angle2, angle1, xOf
 
 	decor := typ.NewDecor()
 	decor.BitSet = typecode
-	decor.Info = typecode2
+	decor.Info = int8(typecode2)
 	decor.X = x*128 + 64 + xOffset
 	decor.Z = z*128 + 64 + zOffset
 	decor.Y = y
@@ -395,7 +395,7 @@ func (w *World3D) AddLoc2(arg0, arg1, arg2, arg3, arg4, arg5, arg6, arg7 int, ar
 	}
 	var22 := typ.NewLocation()
 	var22.BitSet = arg12
-	var22.Info = arg13
+	var22.Info = int8(arg13)
 	var22.Level = arg0
 	var22.X = arg5
 	var22.Z = arg6
@@ -649,17 +649,17 @@ func (w *World3D) GetInfo(arg0, arg1, arg2, arg3 int) int {
 		return -1
 	}
 	if var5.Wall != nil && var5.Wall.BitSet == arg3 {
-		return int(var5.Wall.Info & 0xFF)
+		return int(var5.Wall.Info) & 0xFF
 	}
 	if var5.Decor != nil && var5.Decor.BitSet == arg3 {
-		return int(var5.Decor.Info & 0xFF)
+		return int(var5.Decor.Info) & 0xFF
 	}
 	if var5.GroundDecor != nil && var5.GroundDecor.BitSet == arg3 {
-		return int(var5.GroundDecor.Info & 0xFF)
+		return int(var5.GroundDecor.Info) & 0xFF
 	}
 	for i := range var5.LocCount {
 		if var5.Locs[i].BitSet == arg3 {
-			return int(var5.Locs[i].Info & 0xFF)
+			return int(var5.Locs[i].Info) & 0xFF
 		}
 	}
 	return -1
